requirements.txt
django>=1.8,<1.9

pytz==2016.6.1

django-model-utils==2.5

unicode-slugify==0.1.3
django-autoslug==1.9.3

markdown==2.6.6
django-markup==1.1
